JAPI-34-2 unversioned is latest build.  versioned old interface.
The unversioned interface is now updated to the 5.4.4 version of the
platform, while the previous JAPI was moved to HPCCv500 for backwards
compatibility.  EclRecordParser was moved from the antlr package to the
platform package as it's not platform agnostic.  The eclrecordinfo
information that was stored in the utils.Utils class was moved over to
the platform code, which now makes the entire utils package platform
agnostic. The plan going forward, is to update the generated code on
every major release, which will allow the users of JAPI to get the most
up to date functionality out of the esp.  The unversioned interface will
always be the latest, and whenever we do an update the last platform
release version will be used to name the interface which guarantees
compatibility.